[App-180]: rename spec_config to spec_configuration, set us default RSpec file
