+++ .rspec
@@ -1,1 +1,1 @@
---require spec_config
+--require spec_configuration
